Added possibility to pass the HexData directly
For now, eos-go does not have proper free-form ABI serializer (in progress).
So, for now, it's needed to actually use the ABIJSONToBin API call to
transform the JSON into binary data.

However, it's not possible to use the binary as-is in an action data because
the Data field was automatically serialized to binary which is not correct
as it must match the actually ABI of the contract for the binary to be
smart contract correct.

As a temporary measure, added possibility in the library to not pass Data
and instead assume and use as-is the HexData field which is expected to
contain the binary data of the action.