Remove `omitempty` for now from GetActionsRequest
The default values for those in `cleos` are respectively `-1` and `-20`.
This is causing problems as 0 is interpreted as empty in Go which get it
removed from the request when used like this and will ultimetaly be replaced
by `-1` and `-20` when they reach the API.

The other solution would have been to use a `*int64` type instead. But
pointer literal is not possible easily in Go. Doing so would have render
the calling site more ugly by forcing to create intermediate variables to
specify those.

For now, we simply removed the `omitempty` clauses and will look for maybe
a better solution in the future.