Replaced custom-built logging with `zap` from Uber. FIXED BUGS too. Read on
Intermitent bug when `d.data` would move under the hood, when reading a
Signature and a PublicKey (we were passing a slice as `Content`).

Made tests not depend on time.Now() and check the serialized bytes.

Logger object doesn't exist anymore, but you can enable logging with:
* EnableDecoderLogging
* EnableEncoderLogging
* EnableABIDecoderLogging
* EnableABIEncoderLogging

This commit boosts performances of
* ABI reading by 2.07x
* ABI writing by 1.57x

Uses uber/zap