Fixed encoding/decoding binary form for public keys
The R1 public key was not handled correctly when encoding/decoding it
to its binary form. The R1 key used 37 bytes of storage while the K1 is
using 33 bytes (both uses 1 byte for the curve ID).

This commit fixes all this by adding proper support for encoding/decoding
of R1 public keys.