Fixed encoding/decoding binary form for public keys (bis)
The R1 public key was not handled correctly when encoding/decoding it
to its binary form. The actual creation of the public key from the string
representation was not handled correctly. There is a 4 bytes checksum at
the end of the content buffer. This is now checked correctly, enabling
correct encoding/decoding of R1 key to binary.

At the same time, we properly handling now the PUB_R1 String representation
which does not prints only zeroes now, but prints the correct string
representation of the R1 key.