Turned unknown key error returned from `EOSIO` API into ErrNotFound error
Some endpoint like `/v1/chain/get_account` returns a body in the form:

```
{
  "code": 500,
  "message": "Internal Service Error",
  "error": {
    "code": 0,
    "name": "exception",
    "what": "unspecified",
    "details": [{
      "message": "unknown key (<... redacted ...>): (0 eos.rex)",
      "file": "http_plugin.cpp",
      "line_number": 589,
      "method": "handle_exception"
    }]
  }
}
```

This is problematic as the API only flag ErrNotFound when the HTTP code is
404. We added a new method `IsUnknownKeyError` on the `APIError` struct
so one can check if the `APIError` respects conditions making it an
unknown key error.

When it's the case, the API will now return `ErrNotFound` error in those
cases instead of the weird and bad `EOSIO` error.