Refactor `NewSetContract` to re-use `NewSetCode` and `NewSetABI` and allow empty ABI file
It's now possible to pass an `ABI` file that is fully empty enabling the
possibility to "unset" an `ABI`. It was not working properly because we try
parse as JSON regardless if there is something or not.

At the same time, factorized `NewSetContract` to re-use `NewSetCode` and
`NewSetABI` instead of deplucation the same code twice.