Added support for int compression type in JSON
It appears that `eosjs` output the `compression` fied in JSON using an
integer. This passes when sent to `nodeos` so it seems it's correctly
supported.

Source code of `nodeos` also seems to implies that is possible to use a
string for the `compression` field (from various comments on `cleos` source
code). Investigation was not full in the sense that it's not clear if the
underlying source supports both.

Since the eos-go implementation was supporting string version, I decided
to simply added support for `uint8` support if string fails with an
unmarshal type error.