WIP Blocks log reading
We hit a road block here where the offset is at the end of the block log
completely instead of being right after the end of the genesis state struct.

The problem lies down to the `eos.UnmarshalBinaryReader` method that we
use to decode the struct, the `GenesisState` struct in this case. The
underlying decoder implementation relies on having the full buffer of
the struct available. So, the `eos.UnmarshalBinaryReader` when it starts
do an `ioutil.ReadAll` call to read the full reader and obtain the buffer.

This is highly problematic in this case because we would read the full
blocks log in memory which in this case, it's not viable. Even when the
full blocks log fit in memory (like in our playground files), it's still
a big problem since the offset is all wrong and we need to start at the
right offset for correctly reading the next data (a block or a block
position)