Fixed BlockTimestamp binary & JSON marshalling
The BlockTimestamp is seralized to binary as a uint32 containing a slot.
A slot represents the count of 500ms that elapsed since block epoch which
is set to 2000-01-01T00:00:00Z. It's **not** a timestamp value, really
something completely different.

We fix the binary encoding/decoding for this type to avoid any problem
with the date being wrong.

Also, the block timestamp seems to always be printed using milliseconds
precision in JSON, so we did the same.

Partially solves eoscanada/eosws#176