Avoid escaping HTML when encoding API body
The default behavior of Golang encoder via `json.Marshal` is to escape
HTMl characters by default with a Unicode sequence.

However, nodeos does not correctly handles those creating a a string with
the Unicode escape sequence instead of interpreting the sequence a
character.

To fix that, we now use the json.Encoder to encode and just before encoding,
we set escape HTML to false.