Fixed ReadSymbol binary decoding and Symbol encoding
A Symbol is encoded using a single 64 bits value. However, the decoder
was instead assuming a precision 8 bits followed by a string encoded
binary data.

In fact, a symbol is encoded using a single 64 bits value, the most
significant byte being the precision and the rest is interpreted as a
SymbolCode.