Fixed bug when precision is `MaxUint8`
The bug was really interesting. When the precision is higher than the amount of digit in
the full string, we are padding it with 0s. But the original check to determine if the
length of the string was lower than precision had an overflow bug!

The check was `if len(str) < int(precision + 1)`, but here, precision is a `uint8` type,
so the +1 was overflowing which resulted in a `uint8` to return to `0`! As such, the string
was not expanded correctly and was later creating an out of the bounds error.