Fixed BaseVariant JSON support
The old behavior was not respecting what `nodeos` does on its side for the JSON format. It
sets a tuple of 2 element, the first one being the type name in `string` type. Net
consequence is that our old model that was only knowing the numeric type id couldn't know
to which type decode to.

To overcome the problem, we introduced `VariantDefinition` and `VariantType`. You now
define explicitely the variant definition and uses this definition to implement on your
special type the concrete implementation of `MarshalJSON`, `UnmarshalJSON`, `UnmarshalBinary`.

It also reduces call site implementation as the code is leaner. Order does matter however,
so definition must be ordered the same as the counterpart native variant type.

The new definition uses reflection to extract which type to create for the actual
definition. This could be further improved in a future PR by caching more aggresively
those costy reflect calls in the definition itself, which is define only once and not
on each unmarshal like right now.