Remodel abidecoder. Muuuccho better.
The problem was the use of `sjson.SetBytes()` that would increase the time
exponentially (or linearly, whatever) with the number of iterations in arrays.

A payload with 50,000 uint64 in an `[]uint64` would take 28 minutes to render.
It now takes 57 milliseconds.

This should also improve massively all binary -> json conversion using an ABI.

# Conflicts:
#	go.mod
#	go.sum